commits

+++ src/main/java/org/ereach/inc/services/hospital/EReachHospitalService.java
@@ -1,8 +1,17 @@
 package org.ereach.inc.services.hospital;
 
+import com.cloudinary.Cloudinary;
+import com.cloudinary.Uploader;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.github.fge.jsonpatch.JsonPatch;
+import com.github.fge.jsonpatch.JsonPatchException;
+import com.github.fge.jsonpatch.JsonPatchOperation;
+import com.github.fge.jsonpatch.ReplaceOperation;
 import lombok.AllArgsConstructor;
 import org.ereach.inc.config.EReachConfig;
 import org.ereach.inc.data.dtos.request.AddressCreationRequest;
+import org.ereach.inc.data.dtos.request.AddressUpdateRequest;
 import org.ereach.inc.data.dtos.request.CreateHospitalRequest;
 import org.ereach.inc.data.dtos.request.UpdateHospitalRequest;
 import org.ereach.inc.data.dtos.response.AddressResponse;
@@ -24,20 +33,21 @@ import org.ereach.inc.services.InMemoryDatabase;
 import org.ereach.inc.services.notifications.EReachNotificationRequest;
 import org.ereach.inc.services.notifications.MailService;
 import org.ereach.inc.services.validators.EmailValidator;
-import org.ereach.inc.services.validators.PasswordValidator;
 import org.ereach.inc.utilities.JWTUtil;
 import org.jetbrains.annotations.NotNull;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
+import java.io.IOException;
+import java.util.*;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static org.ereach.inc.data.models.Role.HOSPITAL_ADMIN;
+import static org.apache.catalina.util.Introspection.getDeclaredFields;
 import static org.ereach.inc.data.models.Role.HOSPITAL;
+import static org.ereach.inc.data.models.Role.HOSPITAL_ADMIN;
+import static org.ereach.inc.utilities.AppUtil.doReplace;
+import static org.ereach.inc.utilities.AppUtil.filterEmptyField;
 import static org.ereach.inc.utilities.Constants.*;
 import static org.ereach.inc.utilities.JWTUtil.extractEmailFrom;
 
@@ -50,10 +60,10 @@ public class EReachHospitalService implements HospitalService {
 	private ModelMapper modelMapper;
 	private MailService mailService;
 	private EmailValidator emailValidator;
-	private PasswordValidator passwordValidator;
 	private AddressService addressService;
 	private InMemoryDatabase inMemoryDatabase;
 	private final EReachConfig config;
+	private ObjectMapper objectMapper;
 	@Override
 	public HospitalResponse registerHospital(@NotNull CreateHospitalRequest hospitalRequest) throws FieldInvalidException, RequestInvalidException {
 		emailValidator.validateEmail(hospitalRequest.getHospitalEmail());
@@ -152,7 +162,61 @@ public class EReachHospitalService implements HospitalService {
 	
 	@Override
 	public HospitalResponse editHospitalProfile(UpdateHospitalRequest hospitalRequest) {
+		String logoUrl = scanForFile(hospitalRequest.getLogo());
+		Optional<Hospital> foundHospital = hospitalRepository.findByHospitalEmail(hospitalRequest.getHospitalEmail());
+		AtomicReference<HospitalResponse> atomicReference = new AtomicReference<>();
+		foundHospital.ifPresentOrElse(hospital -> {
+			JsonPatch updatePatch = buildHospitalUpdatePatch(hospitalRequest);
+			try {
+				HospitalResponse response = applyPatchTo(hospital, updatePatch);
+				response.setLogoCloudUrl(logoUrl);
+				atomicReference.set(response);
+			} catch (JsonPatchException e) {
+				throw new EReachUncheckedBaseException(e);
+			}
+		}, ()->{});
+		return atomicReference.get();
+	}
+	
+	private String scanForFile(MultipartFile logo) {
+		if (!logo.isEmpty())
+			return pushToCloud(logo);
 		return null;
+	}
+	
+	private String pushToCloud(MultipartFile logo) {
+		Cloudinary cloudinary = new Cloudinary();
+		Uploader uploader = cloudinary.uploader();
+		Map<String, Object> map = new HashMap<>();
+		map.put("public_id","e-Reach/hospital/media/"+logo.getName());
+		map.put("api_key",config.getCloudApiKey());
+		map.put("api_secret",config.getCloudApiSecret());
+		map.put("cloud_name",config.getCloudApiName());
+		map.put("secure",true);
+		map.put("resource_type", "auto");
+		try{
+			Map<?,?> response = uploader.upload(logo.getBytes(), map);
+			return response.get("url").toString();
+		}catch (IOException exception){
+			throw new EReachUncheckedBaseException(exception+" File upload failed");
+		}
+	}
+	
+	private HospitalResponse applyPatchTo(Hospital hospital, JsonPatch updatePatch) throws JsonPatchException {
+		JsonNode convertedAddress = objectMapper.convertValue(hospital, JsonNode.class);
+		JsonNode updatedNode = updatePatch.apply(convertedAddress);
+		hospital = objectMapper.convertValue(updatedNode, Hospital.class);
+		hospitalRepository.save(hospital);
+		return modelMapper.map(hospital, HospitalResponse.class) ;
+	}
+	
+	private JsonPatch buildHospitalUpdatePatch(UpdateHospitalRequest hospitalRequest) {
+		List<ReplaceOperation> operations = Arrays.stream(getDeclaredFields(AddressUpdateRequest.class))
+				                                  .filter(field -> filterEmptyField(hospitalRequest, field))
+				                                  .map(field -> doReplace(hospitalRequest, field))
+				                                  .toList();
+		List<JsonPatchOperation> patchOperations = new ArrayList<>(operations);
+		return new JsonPatch(patchOperations);
 	}
 	
 	@Override
@@ -162,7 +226,10 @@ public class EReachHospitalService implements HospitalService {
 	
 	@Override
 	public List<HospitalResponse> getAllHospitals() {
-		return null;
+		return hospitalRepository.findAll()
+							     .stream()
+				                 .map(hospital -> modelMapper.map(hospital, HospitalResponse.class))
+							     .toList();
 	}
 	
 	@Override
@@ -193,9 +260,10 @@ public class EReachHospitalService implements HospitalService {
 	}
 	
 	@Override
-	public void removeHospital(String mail) {
-		hospitalRepository.deleteAll();
-		hospitalRepository.deleteByHospitalEmail(mail);
+	public String removeHospital(String mail) throws RequestInvalidException {
+		if (hospitalRepository.existsByHospitalEmail(mail))
+			hospitalRepository.deleteByHospitalEmail(mail);
+		throw new RequestInvalidException(String.format(HOSPITAL_WITH_EMAIL_DOES_NOT_EXIST, mail));
 	}
 	
 	@Override
